requirements.txt
pdfminer
google-api-python-client 
google-auth-oauthlib
python-docx
pinecone-client
langchain
openai
docx2txt
PyPDF2
python-dotenv
